+++ fairlearn/postprocessing/_threshold_optimizer.py
@@ -15,7 +15,7 @@ import pandas as pd
 import random
 
 from sklearn.base import BaseEstimator, ClassifierMixin
-from sklearn.exceptions import NotFittedError
+from sklearn.utils.validation import check_is_fitted
 
 from fairlearn._input_validation import _validate_and_reformat_input
 from ._constants import (LABEL_KEY, SCORE_KEY, SENSITIVE_FEATURE_KEY, OUTPUT_SEPARATOR,
@@ -28,7 +28,6 @@ DIFFERENT_INPUT_LENGTH_ERROR_MESSAGE = "{} need to be of equal length."
 NON_BINARY_LABELS_ERROR_MESSAGE = "Labels other than 0/1 were provided."
 NOT_SUPPORTED_CONSTRAINTS_ERROR_MESSAGE = "Currently only {} and {} are supported " \
     "constraints.".format(DEMOGRAPHIC_PARITY, EQUALIZED_ODDS)
-PREDICT_BEFORE_FIT_ERROR_MESSAGE = "It is required to call 'fit' before 'predict'."
 MULTIPLE_DATA_COLUMNS_ERROR_MESSAGE = "Post processing currently only supports a single " \
     "column in {}."
 SENSITIVE_FEATURE_NAME_CONFLICT_DETECTED_ERROR_MESSAGE = "A sensitive feature named {} or {} " \
@@ -67,24 +66,12 @@ class ThresholdOptimizer(ClassifierMixin, BaseEstimator):
 
     def __init__(self, *, unconstrained_predictor=None, estimator=None,
                  constraints=DEMOGRAPHIC_PARITY, grid_size=1000, flip=True):
-        if unconstrained_predictor and estimator:
-            raise ValueError(EITHER_PREDICTOR_OR_ESTIMATOR_ERROR_MESSAGE)
-        elif unconstrained_predictor:
-            self._unconstrained_predictor = unconstrained_predictor
-            self._estimator = None
-        elif estimator:
-            self._unconstrained_predictor = None
-            self._estimator = estimator
-        else:
-            raise ValueError(PREDICTOR_OR_ESTIMATOR_REQUIRED_ERROR_MESSAGE)
-
-        self._constraints = constraints
-        if self._constraints not in _SUPPORTED_CONSTRAINTS:
-            raise ValueError(NOT_SUPPORTED_CONSTRAINTS_ERROR_MESSAGE)
-
-        self._grid_size = grid_size
-        self._flip = flip
-        self._post_processed_predictor_by_sensitive_feature = None
+        self.grid_size = grid_size
+        self.flip = flip
+        self.post_processed_predictor_by_sensitive_feature = None
+        self.constraints = constraints
+        self.unconstrained_predictor = unconstrained_predictor
+        self.estimator = estimator
 
     @property
     def constraints(self):
@@ -111,6 +98,14 @@ class ThresholdOptimizer(ClassifierMixin, BaseEstimator):
         :type sensitive_features: currently 1D array as numpy.ndarray, list, pandas.DataFrame,
             or pandas.Series
         """
+        if self.unconstrained_predictor and self.estimator:
+            raise ValueError(EITHER_PREDICTOR_OR_ESTIMATOR_ERROR_MESSAGE)
+        elif not self.unconstrained_predictor and not self.estimator:
+            raise ValueError(PREDICTOR_OR_ESTIMATOR_REQUIRED_ERROR_MESSAGE)
+
+        if self.constraints not in _SUPPORTED_CONSTRAINTS:
+            raise ValueError(NOT_SUPPORTED_CONSTRAINTS_ERROR_MESSAGE)
+
         _, _, sensitive_feature_vector = _validate_and_reformat_input(
             X, y, sensitive_features=sensitive_features, enforce_binary_labels=True)
 
@@ -120,25 +115,25 @@ class ThresholdOptimizer(ClassifierMixin, BaseEstimator):
         else:
             y = [int(y_val) for y_val in y]
 
-        if self._estimator:
+        if self.estimator:
             # train estimator on data first
-            self._validate_estimator()
-            self._estimator.fit(X, y, **kwargs)
-            self._unconstrained_predictor = self._estimator
+            self.estimator_ = self.estimator.fit(X, y, **kwargs)
+        else:
+            self.estimator_ = self.unconstrained_predictor
 
-        scores = self._unconstrained_predictor.predict(X)
+        scores = self.estimator_.predict(X)
         threshold_optimization_method = None
-        if self._constraints == DEMOGRAPHIC_PARITY:
+        if self.constraints == DEMOGRAPHIC_PARITY:
             threshold_optimization_method = \
-                self._threshold_optimization_demographic_parity
-        elif self._constraints == EQUALIZED_ODDS:
+                _threshold_optimization_demographic_parity
+        elif self.constraints == EQUALIZED_ODDS:
             threshold_optimization_method = \
                 self._threshold_optimization_equalized_odds
         else:
             raise ValueError(NOT_SUPPORTED_CONSTRAINTS_ERROR_MESSAGE)
 
         self._post_processed_predictor_by_sensitive_feature = threshold_optimization_method(
-            sensitive_feature_vector, y, scores, self._grid_size, self._flip)
+            sensitive_feature_vector, y, scores, self.grid_size, self.flip)
 
     def predict(self, X, *, sensitive_features, random_state=None):
         """Predict label for each sample in X while taking into account sensitive features.
@@ -156,11 +151,11 @@ class ThresholdOptimizer(ClassifierMixin, BaseEstimator):
         if random_state:
             random.seed(random_state)
 
-        self._validate_post_processed_predictor_is_fitted()
+        check_is_fitted(self)
         _, _, sensitive_feature_vector = _validate_and_reformat_input(
             X, y=None, sensitive_features=sensitive_features, expect_y=False,
             enforce_binary_labels=True)
-        unconstrained_predictions = self._unconstrained_predictor.predict(X)
+        unconstrained_predictions = self.estimator_.predict(X)
 
         positive_probs = _vectorized_prediction(
             self._post_processed_predictor_by_sensitive_feature,
@@ -181,43 +176,39 @@ class ThresholdOptimizer(ClassifierMixin, BaseEstimator):
             of the two numbers in each tuple needs to add up to 1.
         :rtype: numpy.ndarray
         """
-        self._validate_post_processed_predictor_is_fitted()
+        check_is_fitted(self)
         _, _, sensitive_feature_vector = _validate_and_reformat_input(
             X, y=None, sensitive_features=sensitive_features, expect_y=False,
             enforce_binary_labels=True)
         positive_probs = _vectorized_prediction(
             self._post_processed_predictor_by_sensitive_feature, sensitive_feature_vector,
-            self._unconstrained_predictor.predict(X))
+            self.estimator_.predict(X))
         return np.array([[1.0 - p, p] for p in positive_probs])
-
-    def _validate_post_processed_predictor_is_fitted(self):
-        if not self._post_processed_predictor_by_sensitive_feature:
-            raise NotFittedError(PREDICT_BEFORE_FIT_ERROR_MESSAGE)
 
     def _threshold_optimization_demographic_parity(self, sensitive_features, labels, scores,
                                                    grid_size=1000, flip=True):
         """Calculate the selection and error rates for every sensitive feature value.
 
         These calculations are made at different
-        thresholds over the scores. Subsequently weighs each sensitive feature value's error by
-        the frequency of the sensitive feature value in the data. The minimum error point is the
-        selected solution, which is recreated by interpolating between two points on the convex
-        hull of all solutions. Each sensitive feature value has its own predictor in the resulting
-        postprocessed predictor, which requires the sensitive feature value as an input.
+        thresholds over the scores. Subsequently weighs each sensitive feature value's error by the
+        frequency of the sensitive feature value in the data. The minimum error point is the selected
+        solution, which is recreated by interpolating between two points on the convex hull of all
+        solutions. Each sensitive feature value has its own predictor in the resulting postprocessed
+        predictor, which requires the sensitive feature value as an input.
 
-        This method assumes that sensitive_features, labels, and scores are non-empty data
-        structures of equal length, and labels contains only binary labels 0 and 1.
+        This method assumes that sensitive_features, labels, and scores are non-empty data structures
+        of equal length, and labels contains only binary labels 0 and 1.
 
-        :param sensitive_features: the feature data that determines the groups for which the
-            parity constraints are enforced
+        :param sensitive_features: the feature data that determines the groups for which the parity
+            constraints are enforced
         :type sensitive_features: list, numpy.ndarray, pandas.DataFrame, or pandas.Series
         :param labels: the labels of the dataset
         :type labels: list, numpy.ndarray, pandas.DataFrame, or pandas.Series
         :param scores: the scores produced by a predictor's prediction
         :type scores: list, numpy.ndarray, pandas.DataFrame, or pandas.Series
         :param grid_size: The number of ticks on the grid over which we evaluate the curves.
-            A large grid_size means that we approximate the actual curve, so it increases the
-            chance of being very close to the actual best solution.
+            A large grid_size means that we approximate the actual curve, so it increases the chance
+            of being very close to the actual best solution.
         :type grid_size: int
         :param flip: allow flipping to negative weights if it improves accuracy.
         :type flip: bool
@@ -258,37 +249,10 @@ class ThresholdOptimizer(ClassifierMixin, BaseEstimator):
                 _interpolate_curve(roc_convex_hull, 'selection', 'error', 'operation',
                                    self._x_grid)
 
-            # Add up errors for the current group multiplied by the probability of the current
-            # group. This will help us in identifying the minimum overall error.
+            # Add up errors for the current group multiplied by the probability of the current group.
+            # This will help us in identifying the minimum overall error.
             error_given_selection += p_sensitive_feature_value * \
                 self._selection_error_curve[sensitive_feature_value]['error']
-
-            logger.debug(OUTPUT_SEPARATOR)
-            logger.debug("Processing %s", str(sensitive_feature_value))
-            logger.debug(OUTPUT_SEPARATOR)
-            logger.debug("DATA")
-            logger.debug(group)
-            logger.debug("ROC curve: convex")
-            logger.debug(roc_convex_hull)
-
-        # Find minimum error point given that at each point the selection rate for each sensitive
-        # feature value is identical by design.
-        i_best_DP = error_given_selection.idxmin()
-        self._x_best = self._x_grid[i_best_DP]
-
-        # create the solution as interpolation of multiple points with a separate predictor per
-        # sensitive feature value
-        predicted_DP_by_sensitive_feature = {}
-        for sensitive_feature_value in self._selection_error_curve.keys():
-            # For DP we already have the predictor directly without complex interpolation.
-            selection_error_curve_result = self._selection_error_curve[sensitive_feature_value] \
-                .transpose()[i_best_DP]
-            predicted_DP_by_sensitive_feature[sensitive_feature_value] = \
-                InterpolatedPredictor(0, 0,
-                                      selection_error_curve_result.p0,
-                                      selection_error_curve_result.operation0,
-                                      selection_error_curve_result.p1,
-                                      selection_error_curve_result.operation1)
 
         logger.debug(OUTPUT_SEPARATOR)
         logger.debug("From ROC curves")